修复视频美颜的部分bug

+++ 随手记/随手记/Classes/BusinessLogicLayer/ZCImageDataBL.m
@@ -8,16 +8,19 @@
 
 #import "ZCImageDataBL.h"
 #import "ZCImageDataTool.h"
+
 @implementation ZCImageDataBL
 + (NSArray *)readImageDataWithPage:(int)page {
     
    return  [ZCImageDataTool readImageDataWithPage:page];
     
 }
+
 + (BOOL)deletImageDataWithArray:(NSArray *)imageDataArray {
     
     return [ZCImageDataTool deletImageDataWithArray:imageDataArray];
 }
+
 +(BOOL)saveImageData:(UIImage *)image {
     
     //调用数据持久层保存数据，注意这里要获取系统时间作为图片的名字，系统时间，图片名字应该在这里获取，因为这属于逻辑层的业务，不应该交给数据持久层来做。
@@ -38,11 +41,12 @@
     NSInteger minute = [comps minute];
     NSInteger second = [comps second];
     
-    NSString *imageNameStr = [NSString stringWithFormat:@"%d/%d/%d/%d/%d/%d",year,month,day,hour,minute,second];
+    NSString *imageNameStr = [NSString stringWithFormat:@"%ld/%ld/%ld/%ld/%ld/%ld",(long)year,month,day,hour,minute,(long)second];
     
     NSData *imageData = UIImageJPEGRepresentation(image, 1.0);
     
     if ([ZCImageDataTool saveImageWithData:imageData imageName:imageNameStr]) {
+        
         return YES;
         
     }else{
